Port zarr-java backend into PyramidBackend abstraction
Port the zarr-java loading code out of ZarrJavaBackedPyramidal5DImageData
into a new ZarrJavaPyramidBackend that implements PyramidBackend<T, V>.
Delete the old ZarrJavaBackedPyramidal5DImageData class and update its
test, ZarrOpenActions, and BackendBenchmark to drive the zarr-java path
through DefaultPyramidal5DImageData plus a ZarrJavaPyramidBackend
instance. Behavior is unchanged.
